+++ lib/src/inspector_controller.dart
@@ -2,6 +2,7 @@ import 'dart:convert';
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:requests_inspector/src/shake.dart';
+import 'package:requests_inspector/src/stopper_filter.dart';
 import 'package:share_plus/share_plus.dart';
 
 import '../requests_inspector.dart';
@@ -10,6 +11,7 @@ import 'har_generator.dart';
 import 'json_pretty_converter.dart';
 import 'helpers/inspector_helper.dart';
 import 'enums/share_type_enum.dart';
+import 'requests_filter.dart';
 
 typedef StoppingRequestCallback = Future<RequestDetails?> Function(
     RequestDetails requestDetails);
@@ -25,6 +27,8 @@ class InspectorController extends ChangeNotifier {
     StoppingRequestCallback? onStoppingRequest,
     StoppingResponseCallback? onStoppingResponse,
     bool defaultTreeViewEnabled = true,
+    bool defaultExpandChildren = true,
+    bool defaultIsDarkMode = true,
   }) =>
       _singleton ??= InspectorController._internal(
         enabled: enabled,
@@ -32,6 +36,8 @@ class InspectorController extends ChangeNotifier {
         onStoppingRequest: onStoppingRequest,
         onStoppingResponse: onStoppingResponse,
         defaultTreeViewEnabled: defaultTreeViewEnabled,
+        defaultExpandChildren: defaultExpandChildren,
+        defaultIsDarkMode: defaultIsDarkMode,
       );
 
   InspectorController._internal({
@@ -40,10 +46,14 @@ class InspectorController extends ChangeNotifier {
     StoppingRequestCallback? onStoppingRequest,
     StoppingResponseCallback? onStoppingResponse,
     required bool defaultTreeViewEnabled,
+    required bool defaultExpandChildren,
+    required bool defaultIsDarkMode,
   })  : _enabled = enabled,
         _showInspectorOn = showInspectorOn,
         _onStoppingRequest = onStoppingRequest,
         _isTreeView = defaultTreeViewEnabled,
+        _expandChildren = defaultExpandChildren,
+        _isDarkMode = defaultIsDarkMode,
         _onStoppingResponse = onStoppingResponse {
     if (_enabled && _allowShaking)
       _shakeDetector = ShakeDetector.autoStart(
@@ -71,8 +81,9 @@ class InspectorController extends ChangeNotifier {
   int _selectedTab = 0;
   bool _requestStopperEnabled = false;
   bool _responseStopperEnabled = false;
-  bool _isDarkMode = true;
-  bool _isTreeView = true;
+  bool _isDarkMode;
+  bool _isTreeView;
+  bool _expandChildren;
 
   final _requestsList = <RequestDetails>[];
   RequestDetails? _selectedRequest;
@@ -81,6 +92,36 @@ class InspectorController extends ChangeNotifier {
   String _searchQuery = '';
   int _totalMatches = 0;
   int _currentMatchIndex = -1;
+
+  // Search & Filters state
+  String _searchUrlQuery = '';
+  RequestMethod? _filterRequestMethod;
+  int? _filterStatusCode;
+
+  // ------------------------------
+
+  // Stoppers Filter State
+  /// TODO: Would be better if we used the [RequestStopperFilter] and [ResponseStopperFilter] classes instead of these separate fields.
+  RequestMethod? _requestStopperFilterMethod;
+  String? _requestStopperFilterUrl;
+  int? _responseStopperFilterStatusCode;
+  String? _responseStopperFilterUrl;
+  // ------------------------------
+
+  RequestMethod? get requestStopperFilterMethod => _requestStopperFilterMethod;
+  String? get requestStopperFilterUrl => _requestStopperFilterUrl;
+  int? get responseStopperFilterStatusCode => _responseStopperFilterStatusCode;
+  String? get responseStopperFilterUrl => _responseStopperFilterUrl;
+
+  bool get hasRequestStopperFilters =>
+      _requestStopperFilterMethod != null ||
+      (_requestStopperFilterUrl != null &&
+          _requestStopperFilterUrl!.trim().isNotEmpty);
+
+  bool get hasResponseStopperFilters =>
+      _responseStopperFilterStatusCode != null ||
+      (_responseStopperFilterUrl != null &&
+          _responseStopperFilterUrl!.trim().isNotEmpty);
 
   int get selectedTab => _selectedTab;
 
@@ -91,6 +132,8 @@ class InspectorController extends ChangeNotifier {
   bool get isDarkMode => _isDarkMode;
 
   bool get isTreeView => _isTreeView;
+
+  bool get expandChildren => _expandChildren;
 
   List<RequestDetails> get requestsList => _requestsList;
 
@@ -103,6 +146,35 @@ class InspectorController extends ChangeNotifier {
   int get totalMatches => _totalMatches;
 
   int get currentMatchIndex => _currentMatchIndex;
+
+  String get searchUrlQuery => _searchUrlQuery;
+
+  RequestMethod? get filterRequestMethod => _filterRequestMethod;
+
+  int? get filterStatusCode => _filterStatusCode;
+
+  bool get areAnyFiltersApplied =>
+      searchUrlQuery.trim().isNotEmpty ||
+      filterRequestMethod != null ||
+      filterStatusCode != null;
+
+  // Computed filtered + searched list
+  List<RequestDetails> get filteredRequestsList {
+    Iterable<RequestDetails> list = [..._requestsList];
+
+    if (_filterRequestMethod != null)
+      list =
+          list.where(RequestMethodFilter(_filterRequestMethod!).requestFilter);
+
+    if (_filterStatusCode != null)
+      list =
+          list.where(RequestStatusCodeFilter(_filterStatusCode!).requestFilter);
+
+    if (_searchUrlQuery.trim().isNotEmpty)
+      list = list.where(RequestUrlFilter(_searchUrlQuery).requestFilter);
+
+    return list.toList(growable: false);
+  }
 
   bool get _allowShaking => [
         ShowInspectorOn.Shaking,
@@ -133,6 +205,97 @@ class InspectorController extends ChangeNotifier {
     _selectedTab = 1;
     _updateTotalMatches();
     notifyListeners();
+  }
+
+  // setters for search & filters
+  void searchForRequests(String value) {
+    if (_searchUrlQuery == value) return;
+    _searchUrlQuery = value;
+    notifyListeners();
+  }
+
+  void setRequestMethodFilter(RequestMethod? method) {
+    if (_filterRequestMethod == method) return;
+    _filterRequestMethod = method;
+    notifyListeners();
+  }
+
+  void setStatusCodeFilter(int? statusCode) {
+    if (_filterStatusCode == statusCode) return;
+    _filterStatusCode = statusCode;
+    notifyListeners();
+  }
+
+  void clearFilters() {
+    _filterRequestMethod = null;
+    _filterStatusCode = null;
+    notifyListeners();
+  }
+
+  void clearSearch() {
+    if (_searchUrlQuery.isEmpty) return;
+    _searchUrlQuery = '';
+    notifyListeners();
+  }
+
+  void setRequestStopperFilterMethod(RequestMethod? method) {
+    if (_requestStopperFilterMethod == method) return;
+    _requestStopperFilterMethod = method;
+    notifyListeners();
+  }
+
+  void setRequestStopperFilterUrl(String? url) {
+    url = url?.trim();
+    if (url != null && url.isEmpty) {
+      url = null;
+    }
+    if (_requestStopperFilterUrl == url) return;
+    _requestStopperFilterUrl = url;
+    notifyListeners();
+  }
+
+  void setResponseStopperFilterStatusCode(int? statusCode) {
+    if (_responseStopperFilterStatusCode == statusCode) return;
+    _responseStopperFilterStatusCode = statusCode;
+    notifyListeners();
+  }
+
+  void setResponseStopperFilterUrl(String? url) {
+    url = url?.trim();
+    if (url != null && url.isEmpty) {
+      url = null;
+    }
+    if (_responseStopperFilterUrl == url) return;
+    _responseStopperFilterUrl = url;
+    notifyListeners();
+  }
+
+  void clearRequestStopperFilters() {
+    _requestStopperFilterMethod = null;
+    _requestStopperFilterUrl = null;
+    notifyListeners();
+  }
+
+  void clearResponseStopperFilters() {
+    _responseStopperFilterStatusCode = null;
+    _responseStopperFilterUrl = null;
+    notifyListeners();
+  }
+
+  bool shouldStopRequest(RequestDetails requestDetails) {
+    final filter = RequestStopperFilter(
+      requestMethod: _requestStopperFilterMethod,
+      urlPattern: _requestStopperFilterUrl,
+    );
+    return filter.shouldStop(requestDetails);
+  }
+
+  bool shouldStopResponse(ResponseDetails responseDetails) {
+    final filter = ResponseStopperFilter(
+      statusCode: _responseStopperFilterStatusCode,
+      urlPattern: _responseStopperFilterUrl,
+    );
+    return filter.shouldStop(responseDetails);
   }
 
   void showInspector() => pageController.jumpToPage(1);
@@ -371,5 +534,10 @@ class InspectorController extends ChangeNotifier {
     }
 
     return parts.join('\n');
+  }
+
+  void toggleExpandChildren() {
+    _expandChildren = !_expandChildren;
+    notifyListeners();
   }
 }